Add support for namespaced protocols and methods
This changes how some of the internals work.  Almost anywhere where a keyword or
symbol was used to represent a method, it is now resolved into a var. The var
is used instead. This allows the use of namespaced keywords and namespaced symbols
for mocks/stubs. Once namespaces are allowed, the use of aliases can complicate things.
However by resolving to the var aliases are normalized to the full namespace.