Raise exception on bad call-count args. Fixes #11.
When `call-count` or `received?` are invoked with arguments, they
expect that final arg to be a vector of args, and don't explicitly
handle the non-vector case. Rather than coerce args and perpetuate
the confusion, this change elects to throw an IllegalArgumentException
when the given args aren't in the correct form. Furthermore, it adds
a new test case demonstrate efficacy of `call-count` in that form
against Java base collection classes like `ArrayList`.